fix message component error
